[#88] Finished Core Commands (except for GUID, which is semi-peripheral)

+++ receivers/native/generic-core-3/src/main/c++/zscript/execution/ZscriptCommandContext.hpp
@@ -263,7 +263,7 @@ public:
         return zscript;
     }
 
-    int getChannelIndex() {
+    uint8_t getChannelIndex() {
         return parseState->getChannelIndex();
     }
